split variables.h into multiple files
