--- include/dlls.h
@@ -1,25 +1,0 @@
-#ifndef _DLLS_H_
-#define _DLLS_H_
-
-enum DLL_ID {
-    DLL_ANIM      = 0x03,
-    DLL_RACE      = 0x04,
-    DLL_AMSEQ     = 0x05,
-    DLL_AMSFX     = 0x06,
-    DLL_NEWCLOUDS = 0x09,
-    DLL_NEWSTARS  = 0x0A,
-    DLL_NEWLFX    = 0x0B,
-    DLL_MINIC     = 0x0C,
-    DLL_EXPGFX    = 0x0D,
-    DLL_MODGFX    = 0x0E,
-    DLL_PROJGFX   = 0x0F,
-    DLL_SCREENS   = 0x14,
-    DLL_TEXT      = 0x15,
-    DLL_SUBTITLES = 0x16,
-    DLL_WATERFX   = 0x18,
-    DLL_CURVES    = 0x1A,
-    DLL_GPLAY     = 0x1D,
-    DLL_SAVEGAME  = 0x1F
-};
-
-#endif //_DLLS_H_
