+++ include/game/actors/player/player.h
@@ -1,0 +1,2 @@
+extern s32 PlayerPosBuffer_index;
+extern struct Vec3_Int PlayerPosBuffer[60]; //seems to buffer player coords with "timestamp"
