+++ include/game/state/gamebits.h
@@ -1,0 +1,2 @@
+extern s32 gFile_BITTABLE;
+extern s16 gSizeBittable;
