+++ include/sys/fault.h
@@ -1,0 +1,6 @@
+//use for crash screens
+struct ErrString{
+	u32 code1;
+	u32 code2;
+	char* text;
+};
