+++ include/sys/menu/menu.h
@@ -1,0 +1,1 @@
+extern u8 MenuInputDelay; //init'd to 5.
