+++ include/sys/thread.h
@@ -1,0 +1,9 @@
+extern OSThread* __osRunningThread;
+extern OSThread* __osRunQueue;
+// this needs double checking. its address is within gMainThreadStack....
+extern u8 gIdleThreadStack[IDLE_THREAD_SIZE];
+extern OSThread gIdleThread;
+extern OSThread gMainThread;
+extern u64 gMainThreadStack[];        // some sort of data
+extern OSSched *osscheduler_;
+extern s8 osscheduler_stack;
