+++ symbol_addrs.txt
@@ -720,7 +720,7 @@ scales_intro_text = 0xb1ff1231; // type:data
 gametext_error_message = 0xb1ff0c61; // type:data
 intro_text = 0xb1fee51e; // type:data
 sample_text = 0xb1fedf1e; // type:data
-ossceduler_stack = 0x800b09c0; // type:data size:0x1
+osscheduler_stack = 0x800b09c0; // type:data size:0x1
 __Dom2SpeedParam = 0x800cd618; // type:data size:0x74
 __Dom1SpeedParam = 0x800cd5a0; // type:data size:0x74
 joyXSign = 0x800a7fc4; // type:data size:0x4
